+++ CMakeLists_files.cmake
@@ -228,7 +228,6 @@ list (APPEND PUBLIC_HEADER_FILES
   opm/autodiff/TransportSolverTwophaseAd.hpp
   opm/autodiff/WellDensitySegmented.hpp
   opm/autodiff/WellStateFullyImplicitBlackoil.hpp
-  opm/autodiff/WellStateFullyImplicitBlackoilDense.hpp
   opm/autodiff/WellStateFullyImplicitBlackoilSolvent.hpp
   opm/autodiff/SimulatorFullyImplicitBlackoilOutput.hpp
   opm/autodiff/VFPProperties.hpp
